refactor: remove utils index file
