--- apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/utils/index.ts
@@ -1,1 +1,0 @@
-export { getNextStep } from './getNextStep'
